--- src/ui/src/app/store/analysis/actionTypes.js
@@ -1,43 +1,0 @@
-export const GET_RUN_SUMMARY = "GET_RUN_SUMMARY";
-export const GET_RUN_SUMMARY_SUCCESS = "GET_RUN_SUMMARY_SUCCESS";
-export const GET_RUN_SUMMARY_FAILURE = "GET_RUN_SUMMARY_FAILURE";
-export const GET_RUN_SUMMARY_RESET = "GET_RUN_SUMMARY_RESET";
-
-export const GET_AVAILABLE_ATTRIBUTES_TAXONSETS =
-  "GET_AVAILABLE_ATTRIBUTES_TAXONSETS";
-export const GET_AVAILABLE_ATTRIBUTES_TAXONSETS_SUCCESS =
-  "GET_AVAILABLE_ATTRIBUTES_TAXONSETS_SUCCESS";
-export const GET_AVAILABLE_ATTRIBUTES_TAXONSETS_FAILURE =
-  "GET_AVAILABLE_ATTRIBUTES_TAXONSETS_FAILURE";
-export const GET_AVAILABLE_ATTRIBUTES_TAXONSETS_RESET =
-  "GET_AVAILABLE_ATTRIBUTES_TAXONSETS_RESET";
-
-export const GET_COUNTS_BY_TAXON = "GET_COUNTS_BY_TAXON";
-export const GET_COUNTS_BY_TAXON_SUCCESS = "GET_COUNTS_BY_TAXON_SUCCESS";
-export const GET_COUNTS_BY_TAXON_FAILURE = "GET_COUNTS_BY_TAXON_FAILURE";
-export const GET_COUNTS_BY_TAXON_RESET = "GET_COUNTS_BY_TAXON_RESET";
-
-export const GET_CLUSTER_SUMMARY = "GET_CLUSTER_SUMMARY";
-export const GET_CLUSTER_SUMMARY_SUCCESS = "GET_CLUSTER_SUMMARY_SUCCESS";
-export const GET_CLUSTER_SUMMARY_FAILURE = "GET_CLUSTER_SUMMARY_FAILURE";
-export const GET_CLUSTER_SUMMARY_RESET = "GET_CLUSTER_SUMMARY_RESET";
-
-export const GET_ATTRIBUTE_SUMMARY = "GET_ATTRIBUTE_SUMMARY";
-export const GET_ATTRIBUTE_SUMMARY_SUCCESS = "GET_ATTRIBUTE_SUMMARY_SUCCESS";
-export const GET_ATTRIBUTE_SUMMARY_FAILURE = "GET_ATTRIBUTE_SUMMARY_FAILURE";
-export const GET_ATTRIBUTE_SUMMARY_RESET = "GET_ATTRIBUTE_SUMMARY_RESET";
-
-export const GET_CLUSTER_METRICS = "GET_CLUSTER_METRICS";
-export const GET_CLUSTER_METRICS_SUCCESS = "GET_CLUSTER_METRICS_SUCCESS";
-export const GET_CLUSTER_METRICS_FAILURE = "GET_CLUSTER_METRICS_FAILURE";
-export const GET_CLUSTER_METRICS_RESET = "GET_CLUSTER_METRICS_RESET";
-
-export const GET_PAIRWISE_ANALYSIS = "GET_PAIRWISE_ANALYSIS";
-export const GET_PAIRWISE_ANALYSIS_SUCCESS = "GET_PAIRWISE_ANALYSIS_SUCCESS";
-export const GET_PAIRWISE_ANALYSIS_FAILURE = "GET_PAIRWISE_ANALYSIS_FAILURE";
-export const GET_PAIRWISE_ANALYSIS_RESET = "GET_PAIRWISE_ANALYSIS_RESET";
-
-export const GET_PLOT = "GET_PLOT";
-export const GET_PLOT_SUCCESS = "GET_PLOT_SUCCESS";
-export const GET_PLOT_FAILURE = "GET_PLOT_FAILURE";
-export const GET_PLOT_RESET = "GET_PLOT_RESET";
