+++ src/ui/src/app/store/analysis/slices/availableAttributesTaxonsetsSlice.js
@@ -1,4 +1,3 @@
-// store/analysis/availableAttributesTaxonsetsSlice.js
 import { createSlice } from "@reduxjs/toolkit";
 
 const initialState = { data: null, loading: false, error: null };
